+++ CHANGES.md
@@ -13,6 +13,7 @@
 * `Driver::cookie_set_json` and `Driver::cookie_get_json` for storing `JsJson`-enabled structures in a cookie
 * `Driver::history_back()` method invoking `history.back()` on window
 * `html_entities` to ease insertion of uncommon letters and symbols in `dom!` macro
+* `on_load` event
 
 ### Removed